updating...

+++ server/src/services/TaskService.js
@@ -51,11 +51,10 @@ exports.getTaskService = async (req) => {
   }
 };
 // getting all task
+
 exports.getAllTask = async (req) => {
   try {
-    // let email = req.headers["email"];
     let data = await TasksModel.aggregate([
-      // { $match: { email: email } },
       {
         $lookup: {
           from: "users",
@@ -65,8 +64,20 @@ exports.getAllTask = async (req) => {
         },
       },
       {
+        $lookup: {
+          from: "users",
+          localField: "createdBy",
+          foreignField: "_id",
+          as: "createdBy",
+        },
+      },
+      {
         $project: {
           "assignTo._id": 0,
+          "createdBy._id": 0,
+          "createdBy.email": 0,
+          "createdBy.password": 0,
+          "createdBy.mobile": 0,
           "assignTo.password": 0,
           "assignTo.createdDate": 0,
         },
@@ -77,6 +88,59 @@ exports.getAllTask = async (req) => {
     return { status: "fail", message: e };
   }
 };
+
+// exports.getAllTask = async (req) => {
+//   try {
+//     const matchStage = {};
+//     // Optional filtering by createdBy
+//     if (req.query.createdBy) {
+//       matchStage.createdBy = new mongoose.Types.ObjectId(req.query.createdBy);
+//     }
+
+//     // Optional filtering by assignTo (must match any in the array)
+//     if (req.query.assignTo) {
+//       matchStage.assignTo = {
+//         $in: [new mongoose.Types.ObjectId(req.query.assignTo)],
+//       };
+//     }
+
+//     let data = await TasksModel.aggregate([
+//       { $match: matchStage },
+
+//       {
+//         $lookup: {
+//           from: "users",
+//           localField: "assignTo",
+//           foreignField: "_id",
+//           as: "assignTo",
+//         },
+//       },
+//       {
+//         $lookup: {
+//           from: "users",
+//           localField: "createdBy",
+//           foreignField: "_id",
+//           as: "createdBy",
+//         },
+//       },
+//       {
+//         $project: {
+//           "assignTo._id": 0,
+//           "createdBy._id": 0,
+//           "createdBy.email": 0,
+//           "createdBy.password": 0,
+//           "createdBy.mobile": 0,
+//           "assignTo.password": 0,
+//           "assignTo.createdDate": 0,
+//         },
+//       },
+//     ]);
+
+//     return { status: "Success", data: data };
+//   } catch (e) {
+//     return { status: "fail", message: e.toString() };
+//   }
+// };
 
 // geting In Progress task
 exports.getInprogressTaskService = async (req) => {
@@ -152,10 +216,15 @@ exports.updaeTask = async (req) => {
 // Delete Task
 exports.deleteTask = async (req) => {
   try {
-    let id = req.params.id;
-    let Query = { _id: id };
-    let data = await TasksModel.deleteOne(Query);
-    return { status: "success", data: data };
+    if (req.headers["role"] === "admin") {
+      let id = req.params.id;
+      let Query = { _id: id };
+      let data = await TasksModel.deleteOne(Query);
+      return { status: "success", data: data };
+    }
+    else {
+      return { status: "fail", message: "Only admin can delete tasks" };
+    }
   } catch (e) {
     return { status: "fail", message: "Something went wrong" };
   }
